requirements.txt
browser-history
win32crypt
pypiwin32
pycryptodome